move DiskDatabase to separate module
